Plain (Minimal) HTML Document

This RMarkdown output format provides a minimal HTML5 template and minimal features for including CSS and JavaScript files in the output source.

Usage

html_document_plain(
  ...,
  css = "normalize",
  script = NULL,
  highlight = "haddock",
  fig_width = 10,
  fig_height = 7,
  fig_retina = 2,
  keep_md = FALSE,
  dev = "png",
  pandoc_args = NULL,
  extra_dependencies = NULL
)

include_script(head = NULL, before = NULL, after = NULL)

Arguments

...

Additional function arguments to pass to the base R Markdown HTML output formatter html_document_base

css

A list of css files to include in the document's <head>. Include "normalize" in the list of css files to include normalize.css, which provides basic style resetting.

script

A list of .js files to include in the document. Use include_script() to choose where in the HTML the script should be sourced.

highlight

One of the pandoc highlight styles.

fig_width

Default width (in inches) for figures

fig_height

Default height (in inches) for figures

fig_retina

Scaling to perform for retina displays (defaults to 2, which currently works for all widely used retina displays). Set to NULL to prevent retina scaling. Note that this will always be NULL when keep_md is specified (this is because fig_retina relies on outputting HTML directly into the markdown document).

keep_md

Keep the markdown file generated by knitting.

dev

Graphics device to use for figure output (defaults to png)

pandoc_args

Additional command line options to pass to pandoc

extra_dependencies

Additional function arguments to pass to the base R Markdown HTML output formatter html_document_base

head, before, after

A character vector of source files, each to be included in the <head>, or before or after the HTML content in <body>.

Functions

include_script: Helper function for including JS scripts